backend/requirements.txt
alembic==1.13.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
bidict==0.22.1
click==8.1.7
dnspython==2.5.0
email-validator==2.1.0.post1
exceptiongroup==1.1.3
fastapi==0.104.0
greenlet==3.0.2
h11==0.14.0
httptools==0.6.1
idna==3.4
importlib-metadata==7.0.1
importlib-resources==6.1.1
iniconfig==2.0.0
install==1.3.5
loguru==0.7.2
Mako==1.3.0
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
pytest==7.4.3
python-dotenv==1.0.0
python-engineio==4.8.2
python-socketio==5.11.0
PyYAML==6.0.1
redis==5.0.1
simple-websocket==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0
zipp==3.17.0
